refactor : CalculationHistoryAdapter.kt 파엘에서 각각의 클래스 분리
